requirements.txt
astroid==2.3.3
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
future==0.18.2
httplib2==0.14.0
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pycparser==2.19
pylint==2.4.4
python-telegram-bot==12.2.0
requests==2.7.0
six==1.13.0
tornado==6.0.3
typed-ast==1.4.0
urllib3==1.25.7
wrapt==1.11.2
